+++ packages/maker-experimental/src/conversion/toMarketplace/shapeExtractors/ActionTypeShapeExtractor.test.ts
@@ -111,6 +111,8 @@ function createMockRidGenerator(
       apiName: string,
       interfaceTypeApiName: string,
     ) => `interface-prop.${interfaceTypeApiName}.${apiName}` as any,
+    generateIptRidFromSptRid: (sptRid: string) =>
+      sptRid.replace("shared-property-type", "interface-property-type") as any,
     generateStructFieldRid: (propertyApiName: string, apiName: string) =>
       `struct-field.${propertyApiName}.${apiName}` as any,
     generateDatasetLocator: (