+++ packages/create-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/create-app
 
+## 2.20.0
+
+### Patch Changes
+
+- @osdk/generator-utils@2.20.0
+
 ## 2.19.0
 
 ### Patch Changes